requirements.txt
Django>=4.0,<5.0
psycopg2-binary>=2.8
djangorestframework>=3.12
drf-yasg>=1.20
django-filter>=2.4
django-bootstrap4>=21.2
python-dotenv>=0.19